Use newtype wrapper for IntMap Arbitrary instance
This is required because QuickCheck 2.8.2 introduced its own instance
which conflicts with the one defined here. Fixes #3.

Also move QuickCheck code to the test-suite module Test.Arbitrary to
avoid any future QuickCheck dependency issues for those using the library.
