+++ graph-core.cabal
@@ -22,16 +22,15 @@ library
                        safe >=0.3,
                        deepseq >=1.3,
                        vector >=0.10,
-                       QuickCheck >=2.6,
                        mtl >=2.1
   hs-source-dirs:      src
-  ghc-options: -Wall -fno-warn-orphans
+  ghc-options: -Wall
 
 test-suite graph-core-tests
   type:                exitcode-stdio-1.0
   hs-source-dirs:      src
   main-is:             Tests.hs
-  other-modules:       Test.NodeManager, Test.Core, Test.Persistence
+  other-modules:       Test.NodeManager, Test.Core, Test.Persistence, Test.Arbitrary
   build-depends:       base >=4.6 && <5,
                        hashable >=1.2,
                        unordered-containers >=0.2,
