+++ src/Test/Persistence.hs
@@ -5,10 +5,11 @@ import Data.Core.Graph
 import Data.Core.Graph.NodeManager
 import Data.Core.Graph.Persistence
 
+import Test.Arbitrary
 import Test.Framework
 
-prop_persistence :: NodeMap Char -> Graph -> Bool
-prop_persistence nodeMap graph =
+prop_persistence :: TestNodeMap Char -> Graph -> Bool
+prop_persistence (TestNodeMap nodeMap) graph =
     let nodeMgr = initNodeManager nodeMap
         (nodeMgr', graph') = loadGraph (persistGraph nodeMgr graph)
     in (nodeMgr' == nodeMgr && graph == graph')
